Newer kmer caller system.
We don't track rolling kmers.  Instead we just look for diffs and
having found one we compute the local kmer.

Note for now the indel-related kmers are ignored and we only
accumulate matching states. (The other cases exist but are commented
out.)  This gives an idea of effective phred-values for pileups that
are "good", rather than pileups involving the numerous small indels.

So we couldn't use those for recalling from a fastq, but we could use
them to recalibrate a pileup (so calls in relation to other aligned calls).