Improve counting for stats.
All true/false types of data are compared against the same total - the
total number of times that kmer has been seen.  Not just seen in that
context.  I think this is more meaningful as we don't know the context
when just looking at the sequence.

All we have is the kmer, and want to know given this kmer (at this
quality) what's the true likelihood of a mismatch, overcall or undercall?