Add an STR finder and classify into in-str and out-str.
MATCH0 is out-str, MATCH1 is in-str, MATCH2 is both combined.
Similarly for the other type codes.

In-str is defined as STR length is 4 or more and at least 3 repeats.
So ACGACG is not an STR as length 6 but 2 copies only.

Out-str is anything that isn't in-str, although a future improvement
could be making it tristate.

This clearly shows in plots for OVER and UNDER that STRs are a
significant source of errors.  This is true for most platforms,
indicating it is likely not sequencing that is the problem but
alignment.  This may be a useful tool for measuring indel realignment
or graph based aligners.