Add a -N option to disable neighbour-kmer removal on error.
If we have ATGCAG -> ATGTAG then kmer ATGTA for central "G" is
initially recorded correctly, but then later removed again because the
next kmer TGTAG is wrong (was TGCAG).  The logic being we can't
confirm or deny the ATGTA kmer.

While true that the *kmer* is nonsensical, the probability of the
central base being correct given the kmer it is within is a valid
statistic.  So -N permits these kmers to be retained.

(Overall it's of minimal impact.)