Experimental code (commented out) for near-N.
This is an alternative to using a bed file where we have the consensus
with call confidence set (eg -C99) and skip regions within 20 bp of
N calls.

In practice it makes almost no difference to the results, so for HG002
at least we can be confident the consensus is good enough to use.